Update actions footer prescriptions detail view
TODO maybe refactor so that selectedrows and confirmandelete are inside hooks?
